requirements.txt
django-googledrive-storage
